+++ functions/src/backfill.js
@@ -1,4 +1,7 @@
-const functions = require("firebase-functions");
+/* eslint-disable indent */
+const {onDocumentWritten} = require("firebase-functions/v2/firestore");
+const {error, info, debug} = require("firebase-functions/logger");
+
 const admin = require("firebase-admin");
 const config = require("./config.js");
 const createTypesenseClient = require("./createTypesenseClient.js");
@@ -8,17 +11,18 @@ admin.initializeApp({
   credential: admin.credential.applicationDefault(),
 });
 
-const validateBackfillRun = (snapshot) => {
-  if (![true, "true"].includes(snapshot.after.get("trigger"))) {
-    functions.logger.error("Skipping backfill. `trigger: true` key " + `was not found in Firestore document ${config.typesenseBackfillTriggerDocumentInFirestore}.`);
+const validateBackfillRun = (data) => {
+  if (![true, "true"].includes(data.after.get("trigger"))) {
+    error("Skipping backfill. `trigger: true` key " + `was not found in Firestore document ${config.typesenseBackfillTriggerDocumentInFirestore}.`);
     return false;
   }
 
   // Check if there's a collection specific sync setup
-  const collectionsToSync = snapshot.after.get("firestore_collections");
+  const collectionsToSync = data.after.get("firestore_collections");
   if (Array.isArray(collectionsToSync) && !collectionsToSync.includes(config.firestoreCollectionPath)) {
-    functions.logger.error(
-        "Skipping backfill. The `firestore_collections` key in " + `${config.typesenseBackfillTriggerDocumentInFirestore} did not contain collection ${config.firestoreCollectionPath}.`,
+    error(
+      "Skipping backfill. The `firestore_collections` key in " +
+        `${config.typesenseBackfillTriggerDocumentInFirestore} did not contain collection ${config.firestoreCollectionPath}. ${collectionsToSync}`,
     );
     return false;
   }
@@ -26,22 +30,36 @@ const validateBackfillRun = (snapshot) => {
   return true;
 };
 
-module.exports = functions.firestore.document(config.typesenseBackfillTriggerDocumentInFirestore).onWrite(async (snapshot, context) => {
-  functions.logger.info(
-      "Backfilling " +
+module.exports = onDocumentWritten("typesense_sync/backfill", async (snapshot, context) => {
+  info(
+    "Backfilling " +
       `${config.firestoreCollectionFields.join(",")} fields in Firestore documents ` +
       `from ${config.firestoreCollectionPath} ` +
       `into Typesense Collection ${config.typesenseCollectionName} ` +
       `on ${config.typesenseHosts.join(",")}`,
   );
 
-  if (!validateBackfillRun(snapshot)) {
+  if (!snapshot.data) {
+    return;
+  }
+
+  if (!validateBackfillRun(snapshot.data)) {
     return false;
   }
 
   const typesense = createTypesenseClient();
 
-  const querySnapshot = await admin.firestore().collection(config.firestoreCollectionPath);
+  const pathSegments = config.firestoreCollectionPath.split("/").filter(Boolean);
+  const pathPlaceholders = utils.parseFirestorePath(config.firestoreCollectionPath);
+  const isGroupQuery = pathSegments.length > 1 && Object.values(pathPlaceholders).length;
+
+  let querySnapshot;
+  if (isGroupQuery) {
+    const collectionGroup = pathSegments.pop();
+    querySnapshot = admin.firestore().collectionGroup(collectionGroup);
+  } else {
+    querySnapshot = admin.firestore().collection(config.firestoreCollectionPath);
+  }
 
   let lastDoc = null;
 
@@ -51,18 +69,31 @@ module.exports = functions.firestore.document(config.typesenseBackfillTriggerDoc
     if (thisBatch.empty) {
       break;
     }
-    const currentDocumentsBatch = await Promise.all(
+    const currentDocumentsBatch = (
+      await Promise.all(
         thisBatch.docs.map(async (doc) => {
-          return await utils.typesenseDocumentFromSnapshot(doc);
+          const docPath = doc.ref.path;
+          const pathParams = utils.pathMatchesSelector(docPath, config.firestoreCollectionPath);
+
+          if (!isGroupQuery || (isGroupQuery && pathParams !== null)) {
+            const typesenseDocument = await utils.typesenseDocumentFromSnapshot(doc, pathParams);
+            if (config.shouldLogTypesenseInserts) {
+              debug(`Backfilling document ${JSON.stringify(typesenseDocument)}`);
+            }
+            return typesenseDocument;
+          } else {
+            return null;
+          }
         }),
-    );
+      )
+    ).filter((doc) => doc !== null);
 
     lastDoc = thisBatch.docs.at(-1) ?? null;
     try {
-      await typesense.collections(encodeURIComponent(config.typesenseCollectionName)).documents().import(currentDocumentsBatch);
-      functions.logger.info(`Imported ${currentDocumentsBatch.length} documents into Typesense`);
+      await typesense.collections(encodeURIComponent(config.typesenseCollectionName)).documents().import(currentDocumentsBatch, {action: "upsert"});
+      info(`Imported ${currentDocumentsBatch.length} documents into Typesense`);
     } catch (error) {
-      functions.logger.error(`Import error in a batch of documents from ${currentDocumentsBatch[0].id} to ${lastDoc.id}`, error);
+      error(`Import error in a batch of documents from ${currentDocumentsBatch[0].id} to ${lastDoc.id}`, error);
       if ("importResults" in error) {
         logImportErrors(error.importResults);
       }
@@ -76,7 +107,7 @@ module.exports = functions.firestore.document(config.typesenseBackfillTriggerDoc
     await new Promise((resolve) => process.nextTick(resolve));
   } while (lastDoc);
 
-  functions.logger.info("Done backfilling to Typesense from Firestore");
+  info("Done backfilling to Typesense from Firestore");
 });
 
 /**
@@ -87,6 +118,6 @@ function logImportErrors(importResults) {
   importResults.forEach((result) => {
     if (result.success) return;
 
-    functions.logger.error(`Error importing document with error: ${result.error}`, result);
+    error(`Error importing document with error: ${result.error}`, result);
   });
 }