+++ test/utilsWithoutFlattening.spec.js
@@ -19,7 +19,7 @@ describe("Utils", () => {
           },
         };
         const documentSnapshot = test.firestore.makeDocumentSnapshot(data, "id");
-        const result = await typesenseDocumentFromSnapshot(documentSnapshot, []);
+        const result = await typesenseDocumentFromSnapshot(documentSnapshot, {}, []);
         expect(result).toEqual({
           id: "id",
           title: "Title X",
@@ -50,7 +50,7 @@ describe("Utils", () => {
           },
           "id",
         );
-        const result = await typesenseDocumentFromSnapshot(documentSnapshot, ["user.name", "user.address.city", "tags"]);
+        const result = await typesenseDocumentFromSnapshot(documentSnapshot, {}, ["user.name", "user.address.city", "tags"]);
         expect(result).toEqual({
           id: "id",
           user: {name: "John Doe", address: {city: "New York"}},
@@ -68,7 +68,7 @@ describe("Utils", () => {
           },
           "id",
         );
-        const result = await typesenseDocumentFromSnapshot(documentSnapshot, ["user.name", "user.address.city"]);
+        const result = await typesenseDocumentFromSnapshot(documentSnapshot, {}, ["user.name", "user.address.city"]);
         expect(result).toEqual({
           id: "id",
           user: {
@@ -89,7 +89,7 @@ describe("Utils", () => {
           },
           "id",
         );
-        const result = await typesenseDocumentFromSnapshot(documentSnapshot, ["title", "user.name"]);
+        const result = await typesenseDocumentFromSnapshot(documentSnapshot, {}, ["title", "user.name"]);
         expect(result).toEqual({
           id: "id",
           title: "Main Title",
@@ -110,7 +110,7 @@ describe("Utils", () => {
           },
           "id",
         );
-        const result = await typesenseDocumentFromSnapshot(documentSnapshot, ["comments.author", "comments.text", "comments.likes"]);
+        const result = await typesenseDocumentFromSnapshot(documentSnapshot, {}, ["comments.author", "comments.text", "comments.likes"]);
         expect(result).toEqual({
           id: "id",
           comments: [